Hide Good Roots from nav, move AvatarMenu into @olivias/ui, trim copy
- Removed /good-roots from the header nav and footer while the app is
  still maturing; the route stays live and is reachable from the avatar
  menu and direct link, but is marked allowIndex:false.
- Promoted AvatarMenu to @olivias/ui with its popover/section/divider
  styles. apps/web now consumes it directly so the component lives
  where other turbo apps can reuse it.
- Supporter tier now reads $10/month.
- Folded the "Every season you get better" section into the "What it
  is" mission band on the Good Roots page.
