+++ apps/web/src/site/routes.ts
@@ -67,11 +67,10 @@ export const routes: AppRoute[] = [
   {
     path: '/good-roots',
     label: 'Good Roots Network',
-    showInNav: true,
-    showInFooter: true,
     title: 'Good Roots Network',
     description: 'A community platform that connects home growers with neighbors and organizations who need fresh food. Plan your garden, see local food gaps, and share what you have extra.',
     seoImage: socialShareImage,
+    allowIndex: false,
     prerender: true,
   },
   {
