+++ packages/ui/src/styles.css
@@ -1501,6 +1501,80 @@ img {
   text-transform: uppercase;
 }
 
+.og-auth-utility__avatar--image {
+  padding: 0;
+  overflow: hidden;
+}
+
+.og-auth-utility__avatar-img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  border-radius: 50%;
+}
+
+.og-auth-menu {
+  position: relative;
+  display: inline-block;
+}
+
+.og-auth-menu .og-auth-utility__avatar {
+  cursor: pointer;
+}
+
+.og-auth-menu__popover {
+  position: absolute;
+  top: calc(100% + 0.5rem);
+  right: 0;
+  min-width: 10rem;
+  padding: 0.35rem;
+  border: 1px solid rgba(79, 56, 37, 0.15);
+  border-radius: 0.75rem;
+  background: var(--og-color-paper);
+  box-shadow: 0 12px 28px rgba(38, 23, 15, 0.14);
+  display: flex;
+  flex-direction: column;
+  gap: 0.15rem;
+  z-index: 40;
+}
+
+.og-auth-menu__item {
+  appearance: none;
+  border: none;
+  background: transparent;
+  text-align: left;
+  padding: 0.55rem 0.85rem;
+  border-radius: 0.5rem;
+  font: inherit;
+  color: var(--og-color-soil);
+  cursor: pointer;
+}
+
+.og-auth-menu__item:hover,
+.og-auth-menu__item:focus-visible {
+  background: rgba(241, 204, 130, 0.28);
+  outline: none;
+}
+
+.og-auth-menu__item--link {
+  display: block;
+  text-decoration: none;
+}
+
+.og-auth-menu__section-label {
+  padding: 0.55rem 0.85rem 0.2rem;
+  font-size: 0.7rem;
+  text-transform: uppercase;
+  letter-spacing: 0.08em;
+  color: rgba(79, 56, 37, 0.6);
+}
+
+.og-auth-menu__divider {
+  height: 1px;
+  margin: 0.35rem 0.5rem;
+  background: rgba(79, 56, 37, 0.12);
+}
+
 /* ── Login / auth page ─────────────────────────────────────────────── */
 
 .og-login-page {
